requirements.txt
asgiref==3.4.1
Django==3.2.9
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-multiselectfield==0.1.12
django-six==1.0.4
pytz==2021.3
screen==1.0.1
sqlparse==0.4.2
xlwt==1.3.0
